wsl: don't set SSH_AUTH_SOCK if it's set already

+++ os-wsl/env.zsh
@@ -12,23 +12,25 @@ if [[ ! -e /.dockerenv ]]; then
             OSTYPE=WSL
             WSLVER=2
             export USERPROFILE="${USERPROFILE:-$(wslpath "$(cmd.exe /D /C 'SET /P <NUL=%USERPROFILE%' 2>/dev/null)")}"
-            if [[ -z "${NPIPERELAY:-}" ]]; then
-                for i in "$USERPROFILE/Apps/npiperelay/npiperelay.exe" "$USERPROFILE/scoop/apps/npiperelay/current/npiperelay.exe"; do
-                    if [[ -e "$i" ]]; then
-                        NPIPERELAY="$i"
-                        break
+            if [[ -z "$SSH_AUTH_SOCK" ]] || [[ "$SSH_AUTH_SOCK" = "$HOME/.ssh/agent.sock" ]]; then
+                if [[ -z "${NPIPERELAY:-}" ]]; then
+                    for i in "$USERPROFILE/Apps/npiperelay/npiperelay.exe" "$USERPROFILE/scoop/apps/npiperelay/current/npiperelay.exe"; do
+                        if [[ -e "$i" ]]; then
+                            NPIPERELAY="$i"
+                            break
+                        fi
+                    done
+                fi
+                if [[ -e "${NPIPERELAY:-}" ]]; then
+                    if (( $+commands[socat] )); then
+                        export SSH_AUTH_SOCK=$HOME/.ssh/agent.sock
+                        if ! ss -a | grep -q "$SSH_AUTH_SOCK"; then
+                            rm -f "$SSH_AUTH_SOCK"
+                            ( setsid socat "UNIX-LISTEN:$SSH_AUTH_SOCK,fork" EXEC:"$NPIPERELAY -ei -s //./pipe/openssh-ssh-agent",nofork & )
+                        fi
+                    else
+                        echo >&2 "Warning: socat is not installed, unable to use npiperelay for ssh auth sock"
                     fi
-                done
-            fi
-            if [[ -e "${NPIPERELAY:-}" ]]; then
-                if (( $+commands[socat] )); then
-                    export SSH_AUTH_SOCK=$HOME/.ssh/agent.sock
-                    if ! ss -a | grep -q "$SSH_AUTH_SOCK"; then
-                        rm -f "$SSH_AUTH_SOCK"
-                        ( setsid socat "UNIX-LISTEN:$SSH_AUTH_SOCK,fork" EXEC:"$NPIPERELAY -ei -s //./pipe/openssh-ssh-agent",nofork & )
-                    fi
-                else
-                    echo >&2 "Warning: socat is not installed, unable to use npiperelay for ssh auth sock"
                 fi
             fi
             ;;
